requirement.txt
django
djangorestframework
deepface
face-recognition
python-dotenv
requests
mtcnn
django-cors-headers
whitenoise==6.6.0
gunicorn